style(Tooltip): clean code
